requirements.txt
boto3
fastapi
mangum
uvicorn
python-dotenv
PyJWT
python-jose